Add a function to merge trivia from a node into an existing trivia
Generally when a node is removed, we need to add its trivia to an
existing node. Add a small helper function to aid in this case.

(cherry picked from commit ca6d6b5341656cbbf51d1b3ec427996e91d26de6)
